requirements.txt
torch
torchvision
librosa
numpy
Pillow